Fix rendering issue and add background image
The commit fixes a rendering issue in the ClassRoomSeatsScreen by setting the initial value of the columnNumper controller to 0 instead of 1. Additionally, the commit adds a new background image to the assets and updates the AssetsManager class to include it. The commit also removes some commented-out code in the ClassRoomController.