Refactor student controller and add edit student dialog
This commit refactors the `StudentController` by adding a list of `StudentResModel` and populating it with data from the server response. It also adds a new import for `multi_dropdown/multiselect_dropdown.dart`. Additionally, it adds a new dialog for editing a student's information, which is displayed when the edit button is pressed in the `StudentScreen`. The `EditStudentWidget` is passed a `StudentResModel` object and populates its fields with the student's information. Some minor changes were made to the `HeaderStudentWidget` to remove a commented out line.