Add functionality to add new exam rooms
Added a new method to the DistributionController class to add new exam rooms. The method sends a POST request to the server with the necessary data and updates the exam rooms list upon success. Also, updated the AddExamRoomWidget to use the new method and added a success snackbar upon successful addition of a new exam room.