Refactor CSV parsing in StudentResModel and StudentController
This commit refactors the CSV parsing in the StudentResModel and StudentController classes. The `fromCsv` method in the StudentResModel class has been removed and replaced with a new `fromCsvWithHeaders` method that takes in a list of headers and returns a StudentResModel object. The StudentController class has been updated to use the new `fromCsvWithHeaders` method and to check for missing headers in the CSV file before parsing. This improves the readability and maintainability of the code.