✨ feat(dio_factory): Remove unused `pretty_dio_logger` package
Removes the `pretty_dio_logger` package from the `dio_factory.dart` file,
as it is no longer being used in the project.

♻️ refactor(cover_widget): Ignore `must_be_immutable` lint rule

Adds the `// ignore: must_be_immutable` comment to the `CoverWidget` class,
as the lint rule is not applicable in this case.

🔧 feat(pubspec): Add `badges` package

Adds the `badges` package to the project's dependencies in the `pubspec.yaml`
file, as it is likely required for a new feature or functionality.