:sparkles: feat: Add BarcodeBindings and remove unused imports
The changes in this commit involve the addition of a new `BarcodeBindings` class and the removal of unused imports from the `bindings.dart` file. The `BarcodeBindings` class is responsible for setting up the dependencies for the `BarcodeController`, which is used in the application.

The removal of the unused imports helps to keep the codebase clean and organized, improving the overall maintainability of the project.