Refactor cover sheets controller and cover widget
This commit refactors the `CoversSheetsController` and `CoverWidget` files. It adds a new variable `schoolTypeId` to the controller and uses it to conditionally render the British and IB cover sheets in the widget. It also updates the UI of the widget by replacing the edit and delete icons with circular avatars and adjusting the width of some elements. Additionally, it removes some commented-out code and unused imports.