✨ refactor(proctor_screen): disable edit proctor functionality
Disables the double-tap functionality to edit a proctor in the ProctorScreen. This change is done to simplify the UI and remove the need for the EditProctorWidget, which is also removed from the file.

Additionally, the childAspectRatio of the GridView is updated to 2 to improve the layout of the exam rooms.