✨ feat: Add proctor list dialog to exam room
Adds a new feature to the Proctor Screen that allows users to view the
list of proctors assigned to a specific exam room. This is triggered
by a double tap on an exam room in the list.

The new `ProctorsInExamRoomWidget` is used to display the list of
proctors within a dialog. This provides a more intuitive way for users
to see which proctors are assigned to each exam room.