🎨 feat(ProctorController): Add method to fetch proctor's exam rooms
This commit adds a new method `getExamRoomsByProctorId` to the `ProctorController`
class. This method is responsible for fetching the exam rooms associated with a
specific proctor. The method makes a GET request to the `examRooms/proctor/{id}`
endpoint and updates the `proctorHasExamRooms` property with the response data.
This change is necessary to support the feature of displaying a list of exam
rooms for a selected proctor.