Generate cover sheet functionality added
Added functionality to generate a cover sheet in the CoversSheetsController. The generateCoverSheet() method sends a GET request to the server to generate a PDF file, and if successful, calls the downloadeTemp() method to download the file. The CoverWidget now includes a button to trigger the generateCoverSheet() method. Also, updated the CoverSeatNumberWidget to display the control mission name and added styling to the widget. Finally, updated the SeatingScreen to display only one CoverSeatNumberWidget at a time and added styling to the widget.