✨ feat: Add ability to set student degree
This change adds the ability to set the student degree in the BarcodeController.
The `setStudentDegree()` method is added, which allows the user to update the
student's degree and persists the changes to the backend. This feature enhances
the functionality of the application by providing a way for users to manage
student degree information.