🎨 feat(assets_manager): add new student degree icon
Adds a new icon for student degrees to the `AssetsManager` class.
The new icon is located at `assets/icons/student_degrees.png` and can
be accessed through the `assetsIconsStudentDegrees` constant.

The changes also include removing the `AssetsManager._()` constructor
and moving the font asset constants to the top of the file for
better organization.