🏫 feat: Implement custom theme in select_school_form widget
This commit introduces a new feature that allows the use of a custom theme
in the `select_school_form` widget. The changes include:

- Importing the `custom_theme/lib.dart` package to access the custom theme
  functionality.
- Removing the unused import of `presentation/resource_manager/index.dart`.

The motivation behind these changes is to provide a more consistent and
visually appealing user experience by integrating the custom theme
throughout the application, starting with the `select_school_form` widget.